Moved package to 'org.freetrm'
Package was incorrectly just 'freetrm'
