--- base/src/freetrm/eventstore/http/EventStoreResponse.scala
@@ -1,7 +1,0 @@
-package freetrm.eventstore.http
-
-
-import freetrm.eventstore.{Event, EventVersionPair}
-import fommil.sjs.FamilyFormats._
-
-
