--- base/src/freetrm/eventstore/utils/Log.scala
@@ -1,7 +1,0 @@
-package freetrm.eventstore.utils
-
-import org.slf4j.LoggerFactory
-
-trait Log {
-  protected val log = LoggerFactory.getLogger(this.getClass)
-}
