+++ base/src/org/freetrm/eventstore/utils/Log.scala
@@ -1,0 +1,7 @@
+package org.freetrm.eventstore.utils
+
+import org.slf4j.LoggerFactory
+
+trait Log {
+  protected val log = LoggerFactory.getLogger(this.getClass)
+}
